+++ ruoyi-modules/ruoyi-system/src/main/java/com/ruoyi/system/service/IChannelService.java
@@ -17,7 +17,7 @@ public interface IChannelService
      * @param id 渠道配置主键
      * @return 渠道配置
      */
-    public Channel selectChannelById(Long id);
+    public Channel selectChannelById(Integer id);
 
     /**
      * 查询渠道配置列表
@@ -33,7 +33,7 @@ public interface IChannelService
      * @param channel 渠道配置
      * @return 结果
      */
-    public Long insertChannel(Channel channel);
+    public int insertChannel(Channel channel);
 
     /**
      * 修改渠道配置
